pyproject.toml
[project]
name = "pitnn"
version = "0.1.0"
description = "Physics-informed tire temperature prediction project scaffold."
readme = "README.md"
authors = [
    { name = "Nicolas Schneider", email = "schneidernicolas90@gmail.com" }
]
requires-python = ">=3.14,<3.15"
dependencies = [
    "fastf1>=3.6.1",
    "hydra-core>=1.3.2",
    "mlx>=0.30.6",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "rich>=14.1.0",
    "tqdm>=4.67.1",
]

[project.scripts]
pitnn = "pitnn:main"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[build-system]
requires = ["uv_build>=0.10.1,<0.11.0"]
build-backend = "uv_build"
